# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "0.1"
    },
    "paths": {
...
        "api/v1/individuals": {
            "get": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Retrive a listing of the Individuals",
                "operationId": "f52ba5ec2f082e5fdefda67b5e0fa722",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Individual"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Store new Individual",
                "operationId": "036a61ac17af2a353e497cf9118de657",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IndividualRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Individual"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/individuals/{uuid}": {
            "get": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Get an existing Individual",
                "operationId": "42c2e979c737ce5a5f9e411036d6ce5b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Individual"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Update an existing Individual",
                "operationId": "96d28c65125a490d82416bea52650bef",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IndividualRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Individual"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Delete existing Individual",
                "operationId": "f9b5e88e9cce4403fe4ce7c7980b2dca",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Individual uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Individual"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IndividualRequest": {
                "title": "IndividualRequest",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "email": {
                        "type": "string",
                        "example": "example"
                    },
                    "address": {
                        "type": "string",
                        "example": "example"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "example"
                    },
                    "vat_number": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
            "Individual": {
                "title": "Individual",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                    },
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "email": {
                        "type": "string",
                        "example": "example"
                    },
                    "address": {
                        "type": "string",
                        "example": "example"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "example"
                    },
                    "vat_number": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
        }
    }
}